Add class
Add DigitalHouseManager

+++ ProfessorAdjunto.kt
@@ -2,7 +2,7 @@ class ProfessorAdjunto(nome:String,
                        sobrenome:String,
                        codigoAdjunto:Int,
                        tempoDeCasa: Float,
-                       var horasMonitoria: Float)
+                       var horasMonitoria: Int)
 
     : Professor(nome, sobrenome,codigoAdjunto, tempoDeCasa) {
 }
